implement textvdf encoder
