--- src/text_vdf/tests/test_encoder.py
@@ -1,1 +1,0 @@
-from .. import TextVdfEncoder, TextVdfDecoder
