+++ src/textvdf/decoder.py
