+++ src/textvdf/tests/vdf_mocks/libraryfolders.vdf
